[SRE] Moved SetMethodBody code to ILGenerator.
